Allow printing non-local type variables
